Update index.html
I was originally defining both a <textarea> and an <input> with the same id="team-collaboration", which causes both elements to display under the same label. The same issue occurs with the id="customer-feedback". 

So, I removed the <input> element from these labels since you already have a <textarea>. You only need one of these for each question, depending on the kind of input you want to allow.